feat(cnivpc-init): implement file copy functionality and version check
- Added a new `copyFile` function to handle file copying with temporary file management.
- Introduced a `checksame` function to compare the versions of source and target binaries before replacement.
- Removed the previous inline file copy implementation from `main.go` for better modularity.
- Updated the deployment configuration to set the `CNI_INIT_OVERWRITE` environment variable.
