+++ cmd/cnivpc-init/main.go
@@ -16,8 +16,8 @@ package main
 
 import (
 	"fmt"
-	"io"
 	"os"
+	"os/exec"
 	"strings"
 
 	"github.com/cockroachdb/errors"
@@ -37,6 +37,14 @@ func main() {
 }
 
 func run() error {
+	same, err := checksame(cniBinarySourcePath, cniBinaryTargetPath)
+	if err != nil {
+		return err
+	}
+	if same {
+		return nil
+	}
+
 	needReplace, err := cniInitModeFromEnv()
 	if err != nil {
 		return err
@@ -46,6 +54,31 @@ func run() error {
 	}
 
 	return installCNIBinaryIfMissing(cniBinarySourcePath, cniBinaryTargetPath)
+}
+
+func checksame(sourcePath, targetPath string) (bool, error) {
+	sourceVersion, err := exec.Command(sourcePath, "--version").CombinedOutput()
+	if err != nil {
+		return false, errors.Wrapf(
+			err,
+			"main.checksame get source version from %s",
+			errors.Safe(sourcePath),
+		)
+	}
+
+	targetVersion, err := exec.Command(targetPath, "--version").CombinedOutput()
+	if errors.Is(err, os.ErrNotExist) {
+		return false, nil
+	}
+	if err != nil {
+		return false, errors.Wrapf(
+			err,
+			"main.checksame get target version from %s",
+			errors.Safe(targetPath),
+		)
+	}
+
+	return strings.TrimSpace(string(sourceVersion)) == strings.TrimSpace(string(targetVersion)), nil
 }
 
 func cniInitModeFromEnv() (bool, error) {
@@ -89,50 +122,5 @@ func copyCNIBinary(sourcePath, targetPath string) error {
 		return errors.Wrap(err, "main.copyCNIBinary copy")
 	}
 	fmt.Printf("Installed CNI binary at %s\n", targetPath)
-	return nil
-}
-
-func copyFile(src, dst string) error {
-	dstTmp := dst + ".tmp"
-	if err := os.Remove(dstTmp); err != nil && !errors.Is(err, os.ErrNotExist) {
-		return errors.Wrapf(err, "main.copyFile remove stale destination %s", errors.Safe(dstTmp))
-	}
-	if err := cp(src, dstTmp); err != nil {
-		return err
-	}
-	if err := os.Rename(dstTmp, dst); err != nil {
-		return errors.Wrapf(err, "main.copyFile rename destination %s", errors.Safe(dst))
-	}
-	return nil
-}
-
-func cp(src, dst string) (err error) {
-	sourceFileInfo, err := os.Stat(src)
-	if err != nil {
-		return errors.Wrapf(err, "main.cp stat source %s", errors.Safe(src))
-	}
-	if !sourceFileInfo.Mode().IsRegular() {
-		return errors.Errorf("main.cp source %s is not a regular file", errors.Safe(src))
-	}
-
-	source, err := os.Open(src)
-	if err != nil {
-		return errors.Wrapf(err, "main.cp open source %s", errors.Safe(src))
-	}
-	defer func() {
-		err = errors.CombineErrors(err, errors.Wrap(source.Close(), "main.cp close source"))
-	}()
-
-	destination, err := os.OpenFile(dst, os.O_WRONLY|os.O_CREATE|os.O_EXCL, sourceFileInfo.Mode().Perm())
-	if err != nil {
-		return errors.Wrapf(err, "main.cp create destination %s", errors.Safe(dst))
-	}
-	defer func() {
-		err = errors.CombineErrors(err, errors.Wrap(destination.Close(), "main.cp close destination"))
-	}()
-
-	if _, err := io.Copy(destination, source); err != nil {
-		return errors.Wrap(err, "main.cp copy")
-	}
 	return nil
 }
